pyproject.toml
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm_pep517_adapter:api" # can be dropped after PDM 1.14
backend-path = ["_buildtool"]

[project]
name = "horsaga.data"
description="Hortensia Saga basic game data and constants"
authors = [{name = "Abel Cheung", email = "abelcheung@gmail.com"}]
requires-python = ">=3.8" # functools.singledispatchmethod
dynamic = ["version"]
license = {text = "MIT"}
keywords = ['hortensia', 'hortensia saga', 'horsaga']
classifiers=[
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Topic :: Games/Entertainment",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: SQL",
]
dependencies = ["attrs"]

[project.urls]
homepage = "https://github.com/Hortensia-Saga/horsaga.data"
Source = "https://github.com/Hortensia-Saga/horsaga.data"

[project.optional-dependencies]
test = [
    "tox-pdm ~= 0.3",
    "pytest",
]
build = [
    "pdm-pep517 ~= 0.9",
]


[tool.pdm]
version = {use_scm = true}
includes = [
    "src",
]
excludes = ["**/*.bak"]
source-includes = [
    "_buildtool/",
    "raw_data/",
    "README.rst",
]

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[tool.pdm.dev-dependencies]
extra = [
    "ptpython",
    "rstcheck",
    "snooty-lextudio",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]